backend/requirements/base.txt
Django==5.2.3
djangorestframework==3.16.0
psycopg[binary]==3.2.9
python-dotenv==1.1.1
django-cors-headers==4.7.0
kavenegar==1.1.2
celery==5.5.3
amqp==5.3.1
django_celery_beat==2.8.1
requests==2.32.4
djangorestframework-simplejwt==5.5.0
django-redis==6.0.0
itsdangerous==2.2.0